+++ ChangeLog.md
@@ -1,7 +1,7 @@
 ### Summary
 
-- Added information about the [Help Center](https://violetwing.featurebase.app/en/help).
-- Completely rewrites the FAQ tab.
-- Displaying information about the latest CS2 patch.
+- Add a new "Offset Source" section in the General settings tab with "Server" and "Local" options.
+- Implement file dialogs for users to specify local `offsets.json`, `client.dll.json`, and `buttons.json` files.
+- Integrate the selected offset source into the application's startup and update logic.
 
 ![Downloads](https://img.shields.io/github/downloads/Jesewe/VioletWing/v1.2.8.3/total?style=for-the-badge&logo=github&color=D5006D) ![Platforms](https://img.shields.io/badge/platform-Windows-blue?style=for-the-badge&color=D5006D) ![License](https://img.shields.io/github/license/jesewe/cs2-triggerbot?style=for-the-badge&color=D5006D)